requirements.txt
Pillow
python-opencv
numpy